filter_cache refactor 1
