+++ src/moderation/__init__.py
@@ -3,6 +3,7 @@ from db import db_m
 from enums import DbPool
 from moderation.user import Permissions, create_user_if_not_exists
 from utils import make_src_path, read_file
+from moderation.filter_cache import get_filter_cache, BaseFilterCache
 
 
 async def init_moderation():
@@ -16,3 +17,6 @@ async def init_moderation():
         admin_password = mod_conf['admin_password']
 
         await create_user_if_not_exists(admin_username, admin_password, True, True, set([p for p in Permissions]), notes=None)
+
+
+fc: BaseFilterCache = get_filter_cache(mod_conf)
